requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
matplotlib
yfinance
scikit-learn
